Separate legacy and SDL keybind formats, allow minimized keybinding files.
Extracted the common filename-to-keybinds path function, at top of file - it's used everywhere.

When reading the available keybindings files, do not list files in the legacy format using SDL DF, and vice versa.

When loading a file, copy if in legacy mode.  If using SDL keybindings, use a serialiser/deserialiser via ordereddict.
This allows pack maintainers and users to store only the relevant parts of a keybinding file, substantially improving
readability and compatibility across versions.

When saving a file in SDL mode, it will use the minified format.